tests: improve test_reference_vmm
Run the tests with using the release build. Instead of doing a cargo
run, we now do it in 2 steps:
1. Build
2. Run

The reason is that with this approach, after step 2 we can just wait
for max 3 seconds for the process to exit. If it did exit, it means that
an error occurred. This wait enables us to print the error to stderr
instead of just failing the test after the test timeout expires.

Signed-off-by: Andreea Florescu <fandree@amazon.com>