make_busybox: remove busybox copy step
If we make changes to make_busybox.sh script, those
changes won't happen because we have a copy step in the
script that will keep the previous busybox (found in
tmp) if the name was not updated. We can remove this
step since the kernel won't be recompiled anyway
(if we don't change something that is related to it).
In the future we might also add a cmd line argument for
the busybox name (-f some_file_path for example).

Signed-off-by: Laura Loghin <lauralg@amazon.com>