tests: add kernel build script test
This test isn't yet run anywhere because it takes forever
(downloads & builds a kernel image for each valid parameter
combination), but it will be included in the long running
test suite once that's set up. Its purpose is to exercise
the (now many) parameters of the kernel image building
script and make sure we don't break anything there.

Signed-off-by: Alexandra Iordache <aghecen@amazon.com>