create a VmState and move config to Vm::new
Instead of making the VMM aware of the configurations that need to
happen when initializing a Vm, just pass all needed state to the
KvmVm::new (for now, it only needs the vcpu count, but it might change
in the future as we add more features).

Because now VmState contains all needed information, we can also just
set mptable right away and remove the call from the VMM.

For all this magic to work, the guest memory configuration is split in
two:
- creating the mmap and the GuestMemory happens in the VMM
- creating the user memory regions happens in the Vm when the Vm is
initialized.

Signed-off-by: Andreea Florescu <fandree@amazon.com>