wrap the IoManager in an Arc<Mutex<>>
A Mutex is not technically required at the moment because we're
doing a one shot configuration of the VMM, and once the Vcpus are
running we no longer add devices. It is added nevertheless because it
makes it much more easier to handle the add of devices transparently on
x8_64 and aarch64, by configuring everything that can be configured in
one shot. This way, VMM::new() returns a configured VMM, and we can
easily add devices without needing to know platform specific details.

For example that the interrupt controller on aarch64 needs to be
created only after creating the vcpus. This detail should not be leaked
from the vm-vcpu crate, and after the point where a VMM is created, we
should be able to add devices at any given point.

This change enables us to do the one shot VMM config. As a consequence,
the VMM TryFrom implementation was removed, and load_kernel and
create_vcpus functions are private and called directly from VMM::new().

Signed-off-by: Andreea Florescu <fandree@amazon.com>