update vcpu to setup the IP only at vcpu.run
The purpose of this change is to allow us to configure create all
resources related to the Vm or Vcpu as soon as we do VMM::new. This is
important because we want as much as possible to avoid making the caller
aware of inner state or functions that need to be called in a
particular order for the VMM to work.

This change is needed because on x86_64 we need to load the
kernel only before running the vcpus because adding devices generates
changes in the kernel command line (as we are using the rudimentary way
of making the guest aware of devices). As the kernel command line can
change from the point it was passed by the user up to the point we are
actually loading it into guest memory, we need to delay that load as
much as possible.

With this change, testing try_from is not possible as it was before.
Added a kernel load test for the happy & error cases instead that covers
the same lines (and a new error path: kernel file exists, but it is
invalid).

Signed-off-by: Andreea Florescu <fandree@amazon.com>