refactor vm-vcpu exit handling
The main purpose of this PR is testability of the vmm & vm-vcpu crates.
In order to easily test both successful cases, but also error cases, we
need to get rid of abrupt exits (via libc::exit).

To replace the exits with a clean shut down which can be tested on
multiple levels of abstractions, the following changes were needed:
- add an exit event handler when initializing the VM. When the vcpus are
exiting from their loop (only in case of errors or on Halt/Shutdown VCPU
exits), we also need to notify the event manager loop that it needs to
exit. For this reason, we're creating a new ExitHandler abstraction
through which we can kick the run loop.
- create a vcpu run loop inside the vcpu::run function so that on Errors
or appropriate vcpu exits (such as Halt), we can just break out of the
loop. With a loop defined inside the Vm::run function, we would need to
introduce another variable to trigger the exit from the loop.

With this new clean shutdown of the VMM, now we can run the rust vmm
integration tests without forking the process, which also simplifies the
test code.

Fixes: https://github.com/rust-vmm/vmm-reference/issues/36

Signed-off-by: Andreea Florescu <fandree@amazon.com>