implement a way to kick vcpus from KVM_RUN
The vcpu threads can now be kicked using real time signals.
The mechanism used is kvm immediate exit + real time signals, which is
adapted from the Crosvm implementation.

The functionality is only used for now to support clean shutdown, but
the plan is to be re-used when adding pause/resume.

Signed-off-by: Andreea Florescu <fandree@amazon.com>