add a run barrier
Make sure that all vcpus start their run loop at the same time. This is
ensured with a run barrier, created by the VM, and sent to each VCPU
as part of their initialization.

On a previous iteration of this commit, the VMM would not longer start.
To catch synchronization problems, the integration tests now run with 4
VCPUs instead of 1.

Signed-off-by: Andreea Florescu <fandree@amazon.com>