Refactoring of args to config parsing
This refactoring aims to deduplicate the parsing effort among the
config's structs and ease the addition of new parameters and structs.

The new `arg_parser` module expose a new `CfgArgParser` struct and
`CfgArgParseError` which will tokenize the input string via `,` first,
and `=` after, and generate an intermediate map rappresentation.

The method `value_of` from the `CfgArgParser` can then be used for
consuming the user-provided parameters.
This method will take care of calling the FromStr's parse function
for converting the input parameter to the required type.

Before exiting the TryFrom of the Configs, the `all_consumed` method
must be called for ensuring no extra parameters are still present in
the tokens map.

The newly introduced CfgArgParseError enum allows for reporting parsing
 errors.

Other then testing for the newly created functions, this test also adds
testing for the conversion of the memory config struct.

Signed-off-by: Federico Ponzi <federico.ponzi92@gmail.com>