tests: improvements to integration tests (#63)
While running the tests inside `test_run_reference_vmm`, at a couple of
places we were running `while True` for a failed test case, that would
cause the test to hang forever, which would get timed out by CI timeout.
Changed this to properly use `subprocess.stdout.read` and fail this if a
test defined timeout is exceeded. This would raise a `TimeoutError` in
such cases.

This is slightly more involved than simply using `Popen.communicate`
or `Popen.wait` alone because, the above closes the `Popen.stdin`, which
we want to be available after we `expect_string` to be able to
`shutdown` the VM later. We set the stdout PIPE to non-blocking and then
check for `BlockingIOError`s to eventually timeout the test cases.

Added test case to make sure that the timeout error is indeed raised

This also fixes #18

Signed-off-by: Abhijit Gadgil <gabhijit@iitbombay.org>