tests: fix long boot time for bzimage-focal (#46)
Since the `bzImage` has got a `RO-RootFS` flag set, the RootFS mounted
was read-only that resulted in a number of startup services to fail and
consequently resulted in longer boot time. Since, we are specifying
separate root, mounting it as `rw` by passing kernel commandline flag
`rw`, the RootFS is mounted as read-write and then the OS boots
normally.

Signed-off-by: Abhijit Gadgil <gabhijit@iitbombay.org>