added tools for downloading resources from S3
The resources in S3 are uploaded using the manifest:
tests/tools/resource_manifest.json.

The download script can download resources based on the
resource name, resource type, and other tags.
Passing the resource name should not be required as filtering can be
done using only the type & tags, but this can be made optional as a
later improvement.

Signed-off-by: Andreea Florescu <fandree@amazon.com>