testing: dld requiered resources from S3
Instead of expecting the required resources for running Rust tests to be
found in /tmp, just use the download script to query for resources that
have the required parameters.

The script for downloading the resources is written in Python. To be
easily to call it from Rust, added a utils crate that has a function for
retrieving the resources. This utilty is only tested with negative
tests to not pollute the test environment.

Signed-off-by: Andreea Florescu <fandree@amazon.com>