tests: add function for downloading resources
Added the download_resource in the s3_download module. This helps with
executing the python integration tests because we no longer need to call
s3_download.py as a script and parse the output. Instead, we just call
the function directly.

While at it, also created some python modules and re-arranged the files.

Signed-off-by: Andreea Florescu <fandree@amazon.com>
Suggested-by: Gabriel Ionescu <gbi@amazon.com>