use debug! macro instead of eprintln
When there is potential of flooding the stdout (i.e. for example in
situations in which the input comes from untrusted sources - guest), use
the debug macro instead of eprintln. With this approach we can still
debug problems, but for production use cases a malicious guest cannot
flood the stdout.

Since the existing Rust dbg! macro is not very helpful in this
scenario, just define an internal debug macro that:
- in debug builds uses eprintl! macro
- in release builds does nothing

Signed-off-by: Andreea Florescu <fandree@amazon.com>