vm-vcpu: refactor create_boot_msr_entries
Return Msrs object instead of a vector of kvm entries. This allows us to
keep in one place the code that creates a static vector with MSRS, and
the code that converts that vector to a FAM struct (`Msrs`). This is
important because we can take the simplifying assumption that the
entries have a valid size, and hence we can safely use unwrap.

Signed-off-by: Andreea Florescu <fandree@amazon.com>