pulled FDT writer code from crosvm
This commit by itself cannot build because of the tests. To keep a clean
history, the tests are fixed in a subsequent commit.

As we are discussing about having vm-fdt a crate in rust-vmm
(see: https://github.com/rust-vmm/community/issues/87), the code is
added to the vm-fdt library crate.

Signed-off-by: Andreea Florescu <fandree@amazon.com>