call vcpu_init & setup GIC
Calling vcpu_init is needed before running the vcpu. This was the
cause of the Exec format error (os error 8).

To see any output from this VM, we need to have a serial console
available. In order to do that we need to:
- implement the MMIO trait for the SerialWrapper (currently only the PIO
trait is implemented as on x86_64 the serial is available on the PIO
bus).
- configure the GIC (interrupt controller) such that we can add the GSI
corresponding to the serial

The code is really chaotic, this is just the bare minimum to start a VM.

Still, there is no output, but the vcpu threads are running.
To be continued...

Signed-off-by: Andreea Florescu <fandree@amazon.com>