use arch defines from firecracker
This makes the VM boot. We need to go back and see why are these values
different from crosvm.

The guest is now stuck because it's doing the time calibration, and for
that it needs to write to 0x3FD, which is a port corresponding to the
RTC device. Next, we need to create & configure the RTC.

Signed-off-by: Andreea Florescu <fandree@amazon.com>