a working arm POC
For this, a few things were needed:
- for setting up the FDT for the serial console and the RTC, use the
Firecracker definitions instead of the Crosvm ones.
- we need to make sure that we register the same GSI in both the FDT,
and using the `register_irqfd` KVM function (this was the reason why the
serial prompt would not show up before)
- setup the RTC to get the proper time.

There are also quite a few things that are not needed, but I just added
because I was unsure why the VM does not boot. These will be explicitly
mentioned in a proper PR.

The RTC implementation from rust-vmm is not correct. Added the
implementation locally, and fixed the function responsible for setting
the correct time in the guest (aka `get_rtc_value`). This will be a fix
in vm-superio.

The code is veeery dirty, the block & net devices will not work because
I hacked the GSI & the MMIO addresses just for the RTC and Serial. This
is a good starting point for figuring out the next steps for adding
support on arm.

A few things that we won't get without:
- abstractions for (in-kernel at least) interrupt controllers
- a resource allocator for MMIO ranges & GSI
- a more comprehensive FDT wrapper, written in a more Rusty way
- unified vcpu abstractions for x86 & aarch64.

To be continued...

Signed-off-by: Andreea Florescu <fandree@amazon.com>