add create_vcpus
This function is needed for aarch64 as creating vcpus one by one and
then just running it is not possible. Between creating the vcpus and
running them, we need to setup the interrupt controller.

To keep a uniform interface between x86_64 & aarch64, we thus create all
vcpus at once, and setup the irq controller for aarch64. This is
transparent for the user. This is also needed so that the end user
doesn't need to call apparently unrelated functions (create vcpu, setup
interrupt controller, and run vcpus) in an order that is not immediately
obvious.

For special use cases, vcpus can be created one by one, but the set up
of the interrupt controller needs to be called manually afterwards.

Some setup (probably in the fdt) is not correct, and running more than
one vcpu yields the following boot error:

[    0.008968] psci: failed to boot CPU1 (-95)
[    0.009427] CPU1: failed to boot: -95

Only one vcpu is successfully running. We need further investigation of
this problem.

Signed-off-by: Andreea Florescu <fandree@amazon.com>