fix tests for aarch64
Some of the tests do not make sense on aarch64 (for example, the ones
related to mptable). There are also a few tests that need to run on
aarch64, but we cannot do that just yet because we do not have an image
on aarch64 that can halt after booting. This will be adressed next.

Signed-off-by: Andreea Florescu <fandree@amazon.com>