update make_test_resources.sh
make_test_resources.sh is now moving the built images
at the same location where the tests are downloading
the needed resources.
Also added a new v4 version that mimics better the folder
structure from vmm-reference, so now it is easier to
upload the resources in S3.
CONTAINER_VERSION file is storing the version of the
container in which the latest images were built.

Signed-off-by: Laura Loghin <lauralg@amazon.com>