save/restore state of the vm & vcpu on x86_64
Call the needed ioctls & return the VmState & VcpuState.
Same for restoring the state.

Still TODO: include the vcpu state in the vm state.

While this support is not implemented completly end to end it does help
us to have it (even partially) here because it helps with delimiting the
interface of the Runnable Vcpu/Runnable VM. It is easier to reason about
what functions need to be public in vm-vcpu-ref in KvmVcpu and KvmVm in
order to support both running the vCPUs and saving their state (through
the RunnableKvm abstractions).

Signed-off-by: Andreea Florescu <fandree@amazon.com>