add RefVirtioDevice trait and struct
These abstractions are used to consolidate functionality which is
common among the virtio devices we use with the reference VMM. The
`RefVirtioDevice` struct implements the common functionality, and
delegates device specific logic to an inner object which represents
a particular device, that is plugged in via the `RefVirtioDeviceT`
trait. Among other things, `RefVirtioDevice` and the environment
traits from the previous commit supersede the `Env` and `CommonConfig`
abstractions which are present right now at the top level in the
`virtio` module (as well as each invidual device implementing traits
and using abstractions from upstream directly). We'll finish the
cleanup after existing devices switch over to the new abstractions.

Signed-off-by: Alexandru Agache <aagch@amazon.com>