cleanup the vm & vcpu interface
The scope is to make it easier to create vcpus from an existing state.
For this goal, a few things where shifted around and changed:
- the vm-vcpu-ref create_vcpus function (part of the VM interface)
  is now private and is automatically called
  when creating the VM. There is no reason for the customers of the
  vm-vcpu crate to manually call create_vcpus when this can be
  automatically called without any impact when the VM is also created.
- deleted the VMM::create_vcpus function

To be able to provide the above mentioned simplifications, the VmConfig
now needs to keep the Vcpu configuration as well. Defined a wrapper over
a list of Vcpu Configurations and added a function to initialize them
with a default value. These are later used by the VMM. Before, the VMM
was creating these defaults, but this is not necessary because the
defaults are pretty "strong" (aka can be used by multiple projects), and
they can be overwritten when they're not appropriate for a project. This
helps with testing the vm-vcpu easier without needing to duplicate that
already existed in the vmm module.

With these changes it is now possible to call the functions that create
the vcpus from an existing state.

Signed-off-by: Andreea Florescu <fandree@amazon.com>