create vm-vcpu-ref crate
This is going to be eventually published to crates.io. This is the crate
that we're using for upstreaming building blocks for setting up the vm
and vcpu for booting.

For now it lives in vmm-reference, but we are considering moving it to
its own worspace at some point in time.

Signed-off-by: Andreea Florescu <fandree@amazon.com>