add rust-vmm-ci to vm-vcpu-ref
This is created as a symlink so we don't need to have another submodule
to maintain. This was we only need to update rust-vmm-ci in one place.

Also, having rust-vmm-ci in vm-vcpu-ref is needed so we can test this as
a separate crate within the workspace. Not all features that we'll add
to vm-vcpu-ref are going to necessarily be used by vmm-reference, so we
could have specialized tests only for vm-vcpu-ref. Also, for publishing
we would like to know what is the coverage of the crate.

For the coverage test to be able to run only for vm-vcpu-ref, we also
need to include this crate in the global workspace.

Signed-off-by: Andreea Florescu <fandree@amazon.com>