update coverage files
From the root coverage file, we exclude the vm-vcpu-ref tests because
these are ran separately. This should make it less likely for both
coverage files to need updates at all times when we're changing
something in vm-vcpu-ref. Also, specify the bindings file to be excluded
using only the file name. This makes it more error prone because as we
change the path of the file, we will still want to ignore it.

For the vm-vcpu-ref, we're adding mpspec.rs to the excluded path.

Signed-off-by: Andreea Florescu <fandree@amazon.com>