use upstream vm-fdt
This is now published to crates.io and we can use it instead of the
local crate. In the second verson of the crate we added a
`property_phandle` function that now can be used to for setting
phandles, thus ensuring their uniqueness.

Also, we should not simply use unwraps when setting up the FDT. We're
now returning an error.

Signed-off-by: Andreea Florescu <fandree@amazon.com>