lapic: return errors for invalid register off
As we're going to export this as the public interface, we should not
assume that callers will pass valid values. Instead of using expects, we
should return errors.

When using this in vm-vcpu for setting up the registers it is safe to
use unwraps because we're using valid values for the register offset.
Returning an error is making the code a bit uglier because the function
where the lapic is set is returning only KvmError, so we would need to
define another error just for the wrong register offset. This would make
the function either return LapicError (the new defined enum variant) and
KvmError.

Signed-off-by: Andreea Florescu <fandree@amazon.com>