[vm-vcpu-ref] upstream gic
This is a working version, but we need a way to retrieve the attributes
used for configuring the GIC. These attributes are different based on
the version of the GIC, and they need to be reused when setting up the
FDT. For that we need to find a good abstraction that does not tie the
Gic to the FDT, as conceptually they're not related.

In this commit we're moving the setup of the interrupt controller on arm
from the VM code (in vm-vcpu), to the vm-vcpu-ref crate. As part of this
move, a few abstractions are implemented in order to achieve the
following goals:
- easy configuration of the number of cpus and number of interrupts
supported by the GIC, as well as the version.
- extensability to support save/restore of the GIC outside this
crate. This is needed because it's likely that the first release of this
crate will not have the support save/restore.

Signed-off-by: Andreea Florescu <fandree@amazon.com>