Adding virito support for arm
for adding support of virtio devices on arm
fdt needs to be updatedd to handle multiple
virtio devices

arm specific changes to EnvMock which is used
to test virtio devices. Incase of arm we first
needed to add a vcpu and register a GIC othervise
register_irqfd call would fail.

`mmio_from_range` function signature is also
changed it now expects reference to `RangeInclusive`
instead of taking its ownership, because we need it
for adding virtio device to fdt.

Closes #210

Signed-off-by: Ramyak mehra <rmehra_be19@thapar.edu>