Update to vm-memory 0.13.1
The new version of vm-memory requires to bump the log crate to version
0.4.20 and vm-virtio to 0.10.0. This has various consequences to the
rest of the code. In particular:

The vmm crate had to be reworked to add a reference to the guest memory
in various places as it is not part of the Queue struct anymore.  For
this reason, it is preferable to wrap the array of Arc<GuestRegionMmap>
with an Arc reference to optimize the various clone() operation that
will be added. This has the advantage of reducing the duplication of
memory and makes sure the various parts of the code to see the same
array of elements (which would not be the case if we clone the array of
Arc<> iteself). While at it, the reference to the guest memory was added
to some key virtio functions.

devices::virtio::{self, block, net} had all to be updated to also accept
the Arc reference to the guest memory.

Signed-off-by: Alvise Rigo <a.rigo@virtualopensystems.com>
Reviewed-by: Timos Ampelikiotis <t.ampelikiotis@virtualopensystems.com>